Colors.h
- Gave names to the struct so that they're not anonymous. Code becomes more standard (eg: rgba.a instead of just a)

DataStorage.cpp

- Changed sz from int to size_t. The value seems to be from filesize which is impossible to be 0. In the return part, it is compared with unsigned int (read_end - read_start), so this prevents the unsigned-signed warning.

entitydata_serializers.cpp and gui_serializers.cpp

- Removed ';' in CEREAL_CLASS_VERSION lines to prevent -Wpedantic error

npc_definitions.h

- Changed idx from size_t (Afaik array/vector/etc indexes is impossible to be 0...)

AuthServer/main.cpp

- Logged reason for shutdown

GameDBSyncEvents.h

- Changed m_current_costume idx from int to size_t
- Removed if condition that checks if m_current_costume_idx >= 0

GameEvents.cpp

- Changed i in the for loop from size_t to i

GameHandler.cpp

- Commented out un-used variable

GameServer.cpp

- Commented out un-used ACE variables

EntityStorage.cpp

- Commented out currently un-used is_incremental

EntityUpdateCodec.cpp

- Commented out unused variables (Those entities are definitely going to be used once the actual functions are done tho)

MapInstance.cpp

- Commented out unused variables

MapServerData.cpp

- Casted m_blah.size() to int
- Should we instead cast level to unsigned int instead? In most games levels are always positive but idk about this game

NetCommandManager.cpp

- Commented out currently-unused commands2

WorldSimulation.cpp

- Don't need to declare float anymore when writing 1.0f (emphasis on the f)
- Commented out vel_scale which is currently unused

bitfield.inl

(Could be a huge minefield created out of this)

- Removed the 'useless' typecastings to glm::uint64 where warnings are given

+++ Projects/CoX/Common/GameData/DataStorage.cpp
@@ -68,7 +68,7 @@ bool BinStore::read_data_blocks( bool file_data_blocks )
         return true;
     }
     QString hdr=read_pstr(20);
-    int sz;
+    size_t sz;
     read_internal(sz);
 
     quint64 read_start = m_str.pos();
