+++ Projects/CoX/Servers/MapServer/MapServerData.cpp
@@ -231,13 +231,13 @@ bool MapServerData::read_runtime_data(const QString &directory_path)
 
 int MapServerData::expForLevel(int lev) const
 {
-    assert(lev>0 && lev<m_experience_and_debt_per_level.m_ExperienceRequired.size());
+    assert(lev>0 && lev<(int)m_experience_and_debt_per_level.m_ExperienceRequired.size());
     return m_experience_and_debt_per_level.m_ExperienceRequired.at(lev - 1);
 }
 
 int MapServerData::expDebtForLevel(int lev) const
 {
-    assert(lev>0 && lev<m_experience_and_debt_per_level.m_DefeatPenalty.size());
+    assert(lev>0 && lev<(int)m_experience_and_debt_per_level.m_DefeatPenalty.size());
     return m_experience_and_debt_per_level.m_DefeatPenalty.at(lev - 1);
 }
